refactor(static code blocks): provide default pragma handler
Pragmas support syntax highlighting and options (such as a summary). Refactor how a pragma is found from a given code node. Much of the plugin’s code was peeled out into smaller, more focused functions. A notable change is that pragma handlers are found by creating a matcher against all known pragmas. This matcher accepts a specification that determines whether a pragma handler is a match for the given code node.

Issue #113